main separation into bot
